LLM/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
redis==5.0.1
faster-whisper>=1.0.0
python-multipart==0.0.9
groq>=0.9.0
